3 - PB/mvp-src/App/requirements.txt
flask
ollama
psycopg2-binary
requests
flask
flask-cors
coverage
scikit-learn
gunicorn
scikit-learn